🚚 Renomme QuestionData::Base en QuestionData
