--- app/models/QuestionData/numeratie.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module QuestionData
-  class Numeratie < QuestionData::Base
-    self.chemin_data = 'questions/numeratie.yml'
-  end
-end
