+++ app/models/question_data.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+class QuestionData < StaticRecord::Base
+  self.chemin_data = 'questions/*.yml'
+
+  attr_reader :nom_technique, :score, :metacompetence
+
+  def initialize(attributes = {})
+    super()
+    @nom_technique = attributes['nom_technique']
+    @score = attributes['score']
+    @metacompetence = attributes['metacompetence']
+  end
+end
