+++ spec/models/question_data_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe QuestionData::Base, type: :model do
+describe QuestionData, type: :model do
   describe 'Chargement de la premiere question' do
     before { stub_static_record_data_loading(yml_data) }
